requirements.txt
# 웹 프레임워크 및 API 관련
fastapi==0.111.0
fastapi-cli==0.0.4
starlette==0.37.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0

# 데이터베이스 관련
alembic==1.13.2
mysql-connector-python==9.0.0
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2

# 데이터 과학 및 머신러닝
numpy==1.25.2
pandas==2.2.2
scikit-learn==1.5.1
scipy==1.13.1
joblib==1.4.2
threadpoolctl==3.5.0
sympy==1.13.0
mxnet==1.7.0.post2
torch==1.10.1
transformers==4.8.1
huggingface-hub==0.0.12
sentencepiece==0.1.96
tokenizers==0.10.3
onnxruntime==1.8.0
tqdm==4.66.4
gluonnlp==0.10.0

# 한국어 처리
konlpy==0.6.0
kobert @ git+https://git@github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3#subdirectory=kobert_hf

# 클라우드 및 AWS 관련
boto3==1.15.18
botocore==1.18.18
s3transfer==0.3.7

# 유틸리티 및 도구
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
Cython==3.0.10
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
graphviz==0.8.4
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
jmespath==0.10.0
JPype1==1.5.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
orjson==3.10.6
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.9
sacremoses==0.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.25.11